WSAS-1817 : supporting webapp stopped status persistence when ghost deployment is enabled
