+++ components/webapp-mgt/org.wso2.carbon.webapp.mgt/src/main/java/org/wso2/carbon/webapp/mgt/WebApplication.java
@@ -31,6 +31,8 @@ import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.CarbonException;
 import org.wso2.carbon.context.PrivilegedCarbonContext;
 import org.wso2.carbon.core.persistence.metadata.ArtifactMetadataException;
+import org.wso2.carbon.registry.core.Registry;
+import org.wso2.carbon.registry.core.exceptions.RegistryException;
 import org.wso2.carbon.tomcat.ext.utils.URLMappingHolder;
 import org.wso2.carbon.utils.CarbonUtils;
 import org.wso2.carbon.utils.FileManipulator;
@@ -424,6 +426,9 @@ public class WebApplication {
             throw new CarbonException("exploded Webapp directory " + webappDir + " deletion failed");
         }
 
+        // remove webapp stopped state from the registry
+        removeWebappStoppedStatus(this);
+
     }
 
     /**
@@ -777,5 +782,26 @@ public class WebApplication {
     private String getWebappKey(){
         return hostName+":"+webappFile.getName();
     }
+
+    /**
+     * Removes webapp stopped entry from the registry
+     *
+     * @param webApplication WebApplication instance
+     */
+    private void removeWebappStoppedStatus(WebApplication webApplication) {
+        if (DataHolder.getRegistryService() != null) {
+            try {
+                Registry configSystemRegistry = DataHolder.getRegistryService().getConfigSystemRegistry(
+                        PrivilegedCarbonContext.getThreadLocalCarbonContext().getTenantId());
+                String webappResourcePath = WebAppUtils.getWebappResourcePath(webApplication);
+                if (configSystemRegistry.resourceExists(webappResourcePath)) {
+                    configSystemRegistry.get(webappResourcePath).removeProperty(WebappsConstants.WEBAPP_STATUS);
+                }
+            } catch (RegistryException e) {
+                log.error("Failed to remove persisted webapp stopped state for: " + webApplication.getContext());
+            }
+        }
+    }
+
 }
